requirements.txt
selenium>=4.15.0
requests>=2.31.0
gspread>=5.12.0
google-auth>=2.23.0
python-dotenv>=1.0.0
